Improving error handling
Summary: Adding failed directories to the transfer report. Also, not reading a
file if it can not be opened. Next step would be to send failure informations
to receiver from sender. Also, receiver has to send list of files failing disk
write to the sender. Will continue to work on this changes.

Reviewed By: @nikunjy

Differential Revision: D1788199