Fixing retry issues, also improving error handling
Summary: 1. Fixing retry issue. A file can get retried from sender while it is
getting written on the receiver side. To fix this, we enqueue the file for
retry only after the next connection has been established. That way, we are
sure that the receiver thread has finished writing.
2. Passing current transfer status information from sender to receiver. Passing
this information in header and with the done command. Information in the header
can be used to trigger fast fail in receiver side. Status information
passed along with done signal will be used by the receiver to determine status
of the entire transfer. Also, passing receiver status information with the done
command. This ensures that sender knows about any disk write or other errors at
receiver.
3. Silently ignoring disk write errors at the receiver side. We still keep
track of the error through error codes and report it back to the sender at the
session end.
4. Added read timeout for client and server sockets. Tested this by not
closing sockets from the other end. Also refactored the socket code to move
common code to SocketUtils.

Reviewed By: @nikunjy

Differential Revision: D1803126