Aborting on file write errors
Summary: When the receiver can't write the file in the directory, then the receiver thread should end the transfer immediately by sending an ABORT command.
After sending the ABORT command the thread should WAIT_FOR_FINISH_WITH_ERROR
It is also possible that the information on the sender side might be wrong becaause we are not making any attempt to send local checkpoints from the failed threads.
This could be done if other threads sent were successful throughout and sent the global checkpoint, but that is not deterministic.

Reviewed By: @uddipta

Differential Revision: D1959902