Syncing disk writes at regular intervals using sync_file_range.
Summary: Maximum value of (read throughput + write throughput) for flash devices
in facebook is 600 mbps. We currently write to flash without waiting for the
content to reach flash from buffer. So, at the transfer end significant amount
of data remains in the buffer. This data can be as much as 20gb and may take
upto 45 sec to sync to flash. Also, we see large drops in throughput. To fix
this, we sync written content using sync_file_range after every 1mb write. We
are using async version of sync_file_range. With this change, we achive constant
write speed of 600mbps. Also, if multiple transfers are running, throughput is
properly distributed among them. This also works for much slower devices like
disk.

Reviewed By: @nikunjy, @ldemailly

Differential Revision: D1969724