Fixing a bug introduced in D1874175
Summary: In the earlier diff, two loops were combined into one. We were trying
to use reference of a vector member, while continuing to call emplace on it.
Emplace invalidates all the member references. wdt_network_test was not failing,
but the logs would show an assertion failure. It did not fail because there was
no diff between src and dst directories. Only the reporting was affected. Fixed
the issue by separating the loops. Also, for assertion failure, started to
inject special string in the log. In wdt_network_test, started to grep for
assertion error to catch such issues in future.

Reviewed By: @ldemailly

Differential Revision: D1984699