Ensuring that e2e simple test can be run in parallel
Summary: e2e simple test was using fixed port. So, it could not be run in
parallel. It now uses dynamic port. Also, during stress testing 48 instances of
the new test is run in parallel. So, 48*4=192 threads read from and write to
disk. This causes extremely slow throughput and pauses. To fix this, ensured
that e2e simple test is using /dev/shm/tmpWDT as the base directory.

Reviewed By: @nikunjy

Differential Revision: D2311269