Fixing network & resumption test failure
Summary: When we block port by dropping packets, connect blocks. And our connect timeout
is 2sec. And even if the port is unblocked, the previous connect does not
succeed and fails after the other side gives up. To fix this, we reduce connect
timeout and try 20 times. This way, we initiate another connect without being
blocked and manage to connect to the server.

Reviewed By: @​#wdt, @ldemailly

Differential Revision: D2437365