Changes to run wdt network and resumption tests in parallel(All tests now run in 7.5 min)
Summary: 1) In network and resumption tests, multiple tests were run and at the end md5s
were verified. But, we had to keep all the snapshots till the very end. That
took too much space. So, now we verify md5 after each test and remove the
destination directory in case of success. In case of failure, the test stops.
2) We were using iptables-save and restore. But, that won't work in case of
simultaneous tests. So, now we remove the particular rule that we added.
3) Both scripts take start_port as a parameter.
4) Resumption tests now run on /dev/shm. We can not run parallel tests on disk.
5) Also, adding version to scuba report

Reviewed By: @​@ldemailly

Differential Revision: D2461868

committer: Service User <svcscm@fb.com>