Fixing a timing bug in download resumption
Summary: There was a timing bug in download resumption.
After file chunk list was sent from receiver to sender, sender would ack it.
But, before doing this, it would set a flag indicating that it has received the
list. In SEND_SETTINGS state, we check this flag and directly start sending
files if this it true. But, the receiver might not have received the ack yet. We
only write transfer log header after we receive the ack. So, a file creation
entry might be written before a header, thereby corrupting the log.
So, removing the optimization of checking for that flag in SEND_SETTINGS state.
Leaving the decision of whether to send chunk list to receiver. This ensures
that file creation entry gets written after the header.

Reviewed By: @nikunjy

Differential Revision: D2491743