Multiple bug fixes related to checkpoint, backward compatibility
Summary: Discovered multiple bugs when I ran network and resumption tests with delay
1) Checkpoint should have seq-id and block offset. Otherwise, we might apply
checkpoint to wrong block.
2) Fixed another bug in readNextCmd in Receiver

Also, tried to run current wdt with the wdt binary deployed. Found another bug
in the way we send local checkpoint. For some reason, we are writing a constant
amount of data when we send local checkpoint. Correct way would be to send size
first. But, that constant has changed before(when we added offset to checkpoint)
and it changes now. This causes a protocol error. To fix this, replaced the
constant with a function with is aware of versions.

If WDT_SENDER or WDT_RECEIVER environment variables are set, network and
resumption test uses those binaries.

Reviewed By: @ldemailly

Differential Revision: D2498069