ensure that data is written to disk, and proper IO error handling
Summary:
there are currently multiple issues in wdt FileWriter that this diff aims to fix:
1/ By default data is not synced to disk at the end of the transfer. This diff ensures that files are always synced to disk
2/ IO errors were almost always silent and would rarely cancel the file transfer. This diff ensures that all IO calls are checked for errors and we fail the transfer on any error.

For some context, zippydb has seen a very high number of corrupted files in the past few months, leading to repeated SEVs and data loss. So far our investigation leads us to think that wdt writer is the culprit here as it doesn't fail the transfer in case of IO errors, which leads to corrupted data on the receiver. I'll try to confirm that by stress testing wdt transfers on a faulty host, with and without this diff.

Reviewed By: uddipta

Differential Revision: D5273341

fbshipit-source-id: 9b906196671058c02effc69751cbe7d454ac7d5a