Remove explicit XCode SDK versions from legocastle jobs
Summary:
Once upon a time there was a lonely hhvm-mac job. It was running merrily through the lego-mac tier. Then someone came and added more hosts to the tier, but something was wrong! They were missing any form of xcode SDK at all! Any job that tried to run in the tier and ended up running on those new hosts failed. Why? Because the jobs were expecting XCode to be present, but it wasn't. To workaround this, an explicit request for xcode was added to the capability list of hhvm-mac. Over time more lego-mac jobs were added, and some gained explicit xcode specifications as well.

The time is long past where explicit version requests for xcode sdks is needed, so let's just remove them. If something starts failing because it's trying to build against `beta`, great, that's useful signal that things are going to break when `beta` becomes mainstream.

Reviewed By: yfeldblum

Differential Revision: D5627221

fbshipit-source-id: c61e359bf44a489f8fc365d78b521c9e2c97ca5e