WDT: Fixing an issue with error processing that may lead to assert
Summary: A legit error status gets overwritten by a shady "error recovery" code. This causes an assert when the sent bytes count mismatch is found. With the fix, the error recovery is bypassed if the byte count check fails. Also, the assert has been removed to avoid crashing a process due to a transient error. (The caller should check the error status and proceed accordingly.)

Reviewed By: nedelchev

Differential Revision: D6517387

fbshipit-source-id: ff47671bcc76eb92a5d4c4c7ee5d86a22733615c