TLS socket bug fix and client change necessary for integration
Summary:
SSLClientSocket::connect can fail without the errno being set to some non-zero value. In that case, we were failing with an assert P60821147. To fix that, we no longer depend on errno.

Also, added support to enable tls through WdtClient api. It gives us an option to override tls values specified in WdtOptions, and needed for wormhole integration.

Differential Revision: D13676677

fbshipit-source-id: 59755dd630ed8586f5ad2a6f8d3fc6cbe2f183ee