FileInfoGenerator for sending files in batches
Summary:
Today sender computes the list of files to be send at the
begining of the transfer, transfers all of the files and
finishes. When we are transferring, for example, a large DB
we might have ended up with new files while the transfer is
in progress, which requires some application side changes to
start another transfer and send these changes. This diff tries
to simplify this by providing the Sender an ability to check
for more files before completing the transfer.

Differential Revision: D17215618

fbshipit-source-id: 3cad04f92bba00efa0510b132c78c3f0989e32ba