Add optional fadvise call following sync_file_range
Summary:
We experimented with enabling DirectIO within FileWriter but the logic
became a bit too complex. The primary goal was to reduce page cache pollution
from files being downloaded if they are not needed immediately.

An alternative approach is to fadvise the cached blocks away periodically
following syncs. Here we add a fadvise call with FAD_DONTNEED immediately
following the call to `sync_file_range` to inform the kernel that the
corresponding pages may be dropped immediately. This call is conditioned on a
new flag added to WdtOptions.

Reviewed By: gwicke

Differential Revision: D19606983

fbshipit-source-id: 2d12639d8c523edd679161990d8c91ea8d212d18